+++ core/camel-support/src/main/java/org/apache/camel/support/builder/BinaryPredicateSupport.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.builder;
+package org.apache.camel.support.builder;
 
 import org.apache.camel.BinaryPredicate;
 import org.apache.camel.Exchange;
